+++ tests/src/rules/default.js
@@ -1,3 +1,4 @@
+import path from 'path'
 import { test, SYNTAX_CASES, getTSParsers } from '../utils'
 import { RuleTester } from 'eslint'
 
@@ -189,6 +190,28 @@ context('TypeScript', function () {
               'import/resolver': { 'eslint-import-resolver-typescript': true },
             },
           }),
+          test({
+            code: `import React from "./typescript-export-assign-default-namespace"`,
+            parser: parser,
+            settings: {
+              'import/parsers': { [parser]: ['.ts'] },
+              'import/resolver': { 'eslint-import-resolver-typescript': true },
+            },
+            parserOptions: {
+              tsconfigRootDir: path.resolve(__dirname, '../../files/typescript-export-assign-default-namespace/'),
+            },
+          }),
+          test({
+            code: `import Foo from "./typescript-export-as-default-namespace"`,
+            parser: parser,
+            settings: {
+              'import/parsers': { [parser]: ['.ts'] },
+              'import/resolver': { 'eslint-import-resolver-typescript': true },
+            },
+            parserOptions: {
+              tsconfigRootDir: path.resolve(__dirname, '../../files/typescript-export-as-default-namespace/'),
+            },
+          }),
         ],
 
         invalid: [
@@ -200,6 +223,24 @@ context('TypeScript', function () {
               'import/resolver': { 'eslint-import-resolver-typescript': true },
             },
             errors: ['No default export found in imported module "./typescript".'],
+          }),
+          test({
+            code: `import React from "./typescript-export-assign-default-namespace"`,
+            parser: parser,
+            settings: {
+              'import/parsers': { [parser]: ['.ts'] },
+              'import/resolver': { 'eslint-import-resolver-typescript': true },
+            },
+            errors: ['No default export found in imported module "./typescript-export-assign-default-namespace".'],
+          }),
+          test({
+            code: `import FooBar from "./typescript-export-as-default-namespace"`,
+            parser: parser,
+            settings: {
+              'import/parsers': { [parser]: ['.ts'] },
+              'import/resolver': { 'eslint-import-resolver-typescript': true },
+            },
+            errors: ['No default export found in imported module "./typescript-export-as-default-namespace".'],
           }),
         ],
       })